Improve startup processes, fix compilation on Linux.
We now use C11, instead of ANSI. Sorry, I want to live in the modern day!
Additionally, various compilation macros have been added so that both Linux
and NetBSD are satisfied and define everything correctly.

We now use realpath when getting files from the command line options, so
that it will be followed correctly later once we daemonize and cd to /

More to come!